refactor: scope each config item
This way, the configuration will be cleaner.
It doesn't have to use the long variable name for each field,
as each of them shares a similar name, such as `url`, `username`, and so on.

Now, each config is scoped, and only the final object is returned, making the code much cleaner.